@@ -7,12 +7,13 @@
   "author": "",
   "license": "",
   "scripts": {
-    "sass": "sass src/styles/app.scss:src/styles/style.css",
-    "watch-sass": "sass --watch src/styles:src/styles",
-    "start": "concurrently \"nodemon --watch src/styles --ext scss --exec 'npm run sass'\" \"NODE_OPTIONS=--openssl-legacy-provider webpack-dev-server --mode development --open\"",
-    "build": "NODE_OPTIONS=--openssl-legacy-provider webpack --mode production"
+    "start": "webpack-dev-server --mode development --open",
+    "build": "NODE_OPTIONS=--openssl-legacy-provider webpack --mode production",
+    "prettier": "prettier --write \"src/**/*.{js,jsx,json,css,scss}\""
   },
   "dependencies": {
+    "axios": "^1.7.2",
+    "date-fns": "^3.6.0",
     "gsap": "^3.12.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -21,17 +22,18 @@
   "devDependencies": {
     "@babel/core": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-react": "^7.24.7",
     "babel-loader": "8.1.0",
     "concurrently": "^8.2.2",
-    "css-loader": "2.1.*",
+    "css-loader": "^7.1.2",
     "html-loader": "0.5.*",
-    "html-webpack-plugin": "3.2.*",
+    "html-webpack-plugin": "^5.6.0",
+    "prettier": "3.3.3",
     "sass": "^1.77.8",
     "sass-loader": "^14.2.1",
-    "style-loader": "0.23.*",
-    "webpack": "4.29.*",
-    "webpack-cli": "3.2.*",
-    "webpack-dev-server": "3.1.*"
+    "style-loader": "^4.0.0",
+    "webpack": "^5.93.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4"
   }
 }
