@@ -1,9 +1,10 @@
 import React from "react";
 
-export default function PluginActionFooterBar() {
+export default function ActionFooterBar() {
   return (
     <div>
       <h1>tab</h1>
     </div>
   );
 }
+
