@@ -1,8 +1,3 @@
-// import './css/body.css';
-// import './css/app.css';
-// import './css/header.css';
-
-import './styles/style.css'
-
-import App from './App.jsx';
+import App from './App.jsx'
+import './styles/app.scss'
 
